refactor: 저수준 모듈에서 고수준 모듈의 Model 사용시 매핑
