+++ app/src/main/java/woowacourse/omok/domain/Game.kt
@@ -1,6 +1,8 @@
 package woowacourse.omok.domain
 
 import woowacourse.omok.adapter.RuleResult
+import woowacourse.omok.domain.event.GameEvent
+import woowacourse.omok.domain.event.PlayEvent
 import woowacourse.omok.domain.model.Board
 import woowacourse.omok.domain.model.rule.OmokRule
 import woowacourse.omok.domain.model.state.OmokState
@@ -8,14 +10,12 @@ import woowacourse.omok.domain.model.state.Turn
 import woowacourse.omok.domain.model.stone.Stone
 import woowacourse.omok.domain.model.stone.Stones
 import woowacourse.omok.domain.repository.StoneRepository
-import woowacourse.ui.GameEvent
-import woowacourse.ui.PlayEvent
 
 class Game(
     private val rule: OmokRule,
     private val stoneRepository: StoneRepository,
     private val gameEvent: GameEvent,
-    board: Board,
+    private val board: Board,
 ) {
     private var state: OmokState = Turn(stoneRepository.lastStoneType(board).reverse())
     private var stones: Stones = stoneRepository.allInBoardSize(board)
@@ -23,7 +23,7 @@ class Game(
     fun initBoard() = gameEvent.initBoard(stones)
 
     fun play(playEvent: PlayEvent) {
-        val position = playEvent.onPosition()
+        val position = playEvent.onPosition(board)
         val stone = Stone(position, state.stoneType)
         val ruleResult = rule.canPlace(stones, stone)
         playEvent.showPlaceResult(ruleResult)
