+++ app/src/main/java/woowacourse/omok/domain/model/position/Position.kt
@@ -1,8 +1,10 @@
 package woowacourse.omok.domain.model.position
 
+import woowacourse.omok.domain.model.Board
+
 data class Position(val column: Column, val row: Row) {
-    constructor(column: Int, row: Int, size: Int) : this(
-        Column.from(column, size),
-        Row.from(row, size),
+    constructor(column: Int, row: Int, board: Board) : this(
+        Column.from(column, board.column),
+        Row.from(row, board.row),
     )
 }
