+++ app/src/main/java/woowacourse/ui/PositionUiModel.kt
@@ -1,0 +1,6 @@
+package woowacourse.ui
+
+data class PositionUiModel(
+    val column: Int,
+    val row: Int,
+)
